Defining the conditional
Not clear of how to do this any other way. We need the ability to define the conditional variable q(x|x_cond), L(x_cond|x) such that for a random walk x_cond = x_initial, and for a Hamiltonian proposal x_cond = -momentum_final. This works but it seems clumsy.